Pack the challenge edge struct, change blocks to uint64 and level to uint8 (#458)

+++ chain-abstraction/sol-implementation/types.go
@@ -102,5 +102,5 @@ type specEdge struct {
 	inner                challengeV2gen.ChallengeEdge
 	startHeight          uint64
 	endHeight            uint64
-	totalChallengeLevels uint64
+	totalChallengeLevels uint8
 }
