+++ contracts/src/challengeV2/IAssertionChain.sol
@@ -19,8 +19,8 @@ interface IAssertionChain {
         bytes32 inboxAcc
     ) external view;
     function validateConfig(bytes32 assertionHash, ConfigData calldata configData) external view;
-    function getFirstChildCreationBlock(bytes32 assertionHash) external view returns (uint256);
-    function getSecondChildCreationBlock(bytes32 assertionHash) external view returns (uint256);
+    function getFirstChildCreationBlock(bytes32 assertionHash) external view returns (uint64);
+    function getSecondChildCreationBlock(bytes32 assertionHash) external view returns (uint64);
     function isFirstChild(bytes32 assertionHash) external view returns (bool);
     function isPending(bytes32 assertionHash) external view returns (bool);
 }
